project2/requirements.txt
python-dotenv
google-auth
google-auth-oauthlib